convert config.py to package
