+++ src/config/__init__.py
@@ -4,130 +4,9 @@ from os.path import join
 
 from src import PROJECT_DIR
 from src.logger import warn
-
-
-class OldBackupSettings:
-    # Number of backups to keep
-    # Only keep when it's not
-    # overdue for deletion.
-    keep: int = 5
-    # If a backup's creation date
-    # exceeds this number, it'll
-    # be deleted.
-    retention: int = 30
-    # Should this program delete
-    # older backups (even if it was
-    # within retention days and
-    # total files less than 'keep')
-    # to save space for new backup.
-    # This option always keeps the most
-    # recent backup unless 'aggressive'
-    # is set to True.
-    del_old_4_space: bool = False
-    # If True, delete to the last backup
-    # to make some space for new backup
-    aggressive: bool = False
-
-    def __init__(self, configuration: dict):
-        #
-        #   Set Keep
-        #
-        keep = configuration['keep']
-        if isinstance(keep, int):
-            self.keep = int(keep)
-        else:
-            warn(f'{keep} is not a valid number for "old_backups.keep!"')
-            warn(f'Use default value: {self.keep}')
-
-        #
-        #   Set Retention
-        #
-        retention = configuration['retention']
-        if isinstance(retention, int):
-            self.retention = int(retention)
-        else:
-            warn(f'{retention} is not a valid number for "old_backups.retention!"')
-            warn(f'Use default value: {self.retention}')
-
-        #
-        #   Set Delete Old Backups For Space Boolean
-        #
-        del_old_4_space = configuration['remove_old_backups_for_space']
-        if isinstance(del_old_4_space, bool):
-            self.del_old_4_space = del_old_4_space
-        else:
-            warn('"old_backups.remove_old_backups_for_space must be a True or False!')
-            warn(f'Use default value: {self.del_old_4_space}')
-
-        #
-        #   Set Aggressive Boolean
-        #
-        aggressive = configuration['aggressive']
-        if isinstance(aggressive, bool):
-            self.aggressive = aggressive
-        else:
-            warn('"old_backups.aggressive must be a True or False!')
-            warn(f'Use default value: {self.aggressive}')
-
-    def __str__(self):
-        return (f'OldBackupSettings('
-                f'keep={self.keep}, '
-                f'retention={self.retention}, '
-                f'del_old_4_space={self.del_old_4_space}, '
-                f'aggressive={self.aggressive}'
-                f')')
-
-
-class ZstdArguments:
-    # Higher values are slower but yield smaller size.
-    # More at: https://python-zstandard.readthedocs.io/en/latest/compressor.html#zstdcompressor
-    level: int = 3
-    # How many threads should the program use.
-    # More threads equals faster compression time.
-    # 0 will use all threads (also the default)
-    threads: int = 0
-
-    def __init__(self, configuration: dict):
-        #
-        #   Set Compression Level
-        #
-        level = configuration['level']
-        if isinstance(level, int):
-            self.level = level
-        else:
-            warn(f'{level} is not a valid number for "arguments.level!"')
-            warn(f'Use default value: {self.level}')
-
-        #
-        #   Set Number of Threads
-        #
-        threads = configuration['threads']
-        if isinstance(threads, int):
-            self.threads = threads
-        else:
-            warn(f'{threads} is not a valid number for "arguments.threads!"')
-            warn(f'Use default value: {self.threads}')
-
-    def __str__(self):
-        return (f'ZstdArgument('
-                f'level={self.level}, '
-                f'threads={self.threads}'
-                f')')
-
-
-class ProgressBar:
-    enabled: bool = False
-
-    def __init__(self, configuration: dict):
-        self.enabled = configuration['enabled']
-
-
-class Settings:
-    write_chunk: int = 1024
-    progress_bar: ProgressBar
-
-    def __init__(self, settings: dict):
-        self.progress_bar = ProgressBar(settings['progress_bar'])
+from .OldBackupsSettings import OldBackupSettings
+from .ZstdArguments import ZstdArguments
+from .settings import Settings
 
 
 class Configuration:
@@ -222,14 +101,16 @@ class Configuration:
         self.settings = Settings(configuration['settings'])
 
     def __str__(self):
-        return (f'Configuration('
-                f'level={self.log_level}, '
-                f'include={str(self.include)}, '
-                f'destination={self.destination}, '
-                f'ignore={str(self.ignore_paths)}, '
-                f'OldBackupSettings={str(self.old_backup_settings)}, '
-                f'ZstdArguments={str(self.zstd_arguments)}'
-                f')')
+        return (
+            f'Configuration('
+            f'level={self.log_level}, '
+            f'include={str(self.include)}, '
+            f'destination={self.destination}, '
+            f'ignore={str(self.ignore_paths)}, '
+            f'OldBackupSettings={str(self.old_backup_settings)}, '
+            f'ZstdArguments={str(self.zstd_arguments)}'
+            f')'
+        )
 
 
 def verify(config: Configuration) -> bool:
