+++ src/config/settings/ProgressBar.py
@@ -1,0 +1,5 @@
+class ProgressBar:
+    enabled: bool = False
+
+    def __init__(self, configuration: dict):
+        self.enabled = configuration['enabled']
